chore: rework ui select
