--- src/ui/select/index.ts
@@ -1,1 +1,0 @@
-export * from './select.component'
