+++ src/ui/select/select-articles-paer-page/index.ts
@@ -1,0 +1,1 @@
+export * from './select-articles-per-page.component'
