+++ src/ui/select/select-articles-paer-page/select-articles-per-page-value.interfaces.ts
@@ -1,0 +1,7 @@
+import { ChangeEventHandler } from 'react'
+
+export interface SelectArticlesPerPageProps {
+  id: string
+  onChange: ChangeEventHandler<HTMLSelectElement>
+  perPageValue: number
+}
