+++ src/ui/select/select-articles-paer-page/select-articles-per-page-value.styles.module.css
@@ -1,14 +1,3 @@
-.selectSort {
-  width: 175px;
-  height: 30px;
-  border: 2px solid var(--primary-color);
-  border-radius: 8px;
-  background-color: var(--second-color);
-  padding: 5px 20px;
-  color: var(--primary-color);
-  font-size: 14px;
-}
-
 .selectPerPage {
   width: 110px;
   height: 30px;
