+++ src/ui/select/select-sort-value/index.ts
@@ -1,0 +1,1 @@
+export * from './select-sort-value.component'
