+++ src/ui/select/select-sort-value/select-sort-value.styles.module.css
@@ -1,0 +1,21 @@
+.selectSort {
+  width: 175px;
+  height: 30px;
+  border: 2px solid var(--primary-color);
+  border-radius: 8px;
+  background-color: var(--second-color);
+  padding: 5px 20px;
+  color: var(--primary-color);
+  font-size: 14px;
+}
+
+.select:focus {
+  outline: none;
+  border: 2px solid var(--primary-color);
+}
+
+.label {
+  font-size: 14px;
+  color: var(--second-color);
+  line-height: 1.2;
+}
